Update create_data.php
Worked on part 4 - 6. Not complete yet. I have it where it displays the different arrays in columns. Below is a table with the random generate data for 25 people. It also only writes on set of data to a file (I need to figure out how to get it to write all 25). Far from finished, but I though I might upload my changes for now.